:recycle: Slight refactors + disable message when not on island
